final_project/requirements.txt
pytest~=7.0.1
selenium~=4.1.3
webdriver-manager~=3.5.3
pytest-xdist~=2.5.0
allure-pytest~=2.9.45
requests~=2.27.1
pymysql~=1.0.2
sqlalchemy~=1.4.35
flask~=2.1.1
Faker~=13.6.0